Refactored the NRLMSIS00 class.
