+++ src/models/atmosphere/CMakeLists.txt
@@ -1,13 +1,15 @@
 set(SOURCES FGMSIS.cpp
-            FGMSISData.cpp
             FGMars.cpp
             FGStandardAtmosphere.cpp
-            FGWinds.cpp)
+            FGWinds.cpp
+            MSIS/nrlmsise-00.c
+            MSIS/nrlmsise-00_data.c)
 
 set(HEADERS FGMSIS.h
             FGMars.h
             FGStandardAtmosphere.h
-            FGWinds.h)
+            FGWinds.h
+            MSIS/nrlmsise-00.h)
 
 add_library(Atmosphere OBJECT ${HEADERS} ${SOURCES})
 set_target_properties(Atmosphere PROPERTIES TARGET_DIRECTORY
