+++ src/models/atmosphere/FGStandardAtmosphere.h
@@ -95,7 +95,7 @@ will be consistently and accurately calculated.
 CLASS DECLARATION
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%*/
 
-class FGStandardAtmosphere : public FGAtmosphere {
+class JSBSIM_API FGStandardAtmosphere : public FGAtmosphere {
 public:
   /// Constructor
   FGStandardAtmosphere(FGFDMExec*);
