+++ src/models/atmosphere/MSIS/nrlmsise-00_data.c
@@ -1,13 +1,9 @@
-// MSIS-00 Data
-// Adapted for use in JSBSim
-// David Culp, davidculp2@comcast.net
-
 /* -------------------------------------------------------------------- */
 /* ---------  N R L M S I S E - 0 0    M O D E L    2 0 0 1  ---------- */
 /* -------------------------------------------------------------------- */
 
 /* This file is part of the NRLMSISE-00  C source code package - release
- * 20020503
+ * 20041227
  *
  * The NRLMSISE-00 model was developed by Mike Picone, Alan Hedin, and
  * Doug Drob. They also wrote a NRLMSISE-00 distribution package in 
@@ -15,7 +11,7 @@
  * http://uap-www.nrl.navy.mil/models_web/msis/msis_home.htm
  *
  * Dominik Brodowski implemented and maintains this C version. You can
- * reach him at devel@brodo.de. See the file "DOCUMENTATION" for details,
+ * reach him at mail@brodo.de. See the file "DOCUMENTATION" for details,
  * and check http://www.brodo.de/english/pub/nrlmsise/index.html for
  * updated releases of this package.
  */
@@ -25,8 +21,6 @@
 /* ------------------------------------------------------------------- */
 /* ------------------------ BLOCK DATA GTD7BK ------------------------ */
 /* ------------------------------------------------------------------- */
-
-namespace JSBSim {
 
 /* TEMPERATURE */
 double pt[150] = {
@@ -744,4 +738,3 @@ double pavgm[10] = {
      2.61000E+02, 2.64000E+02, 2.29000E+02, 2.17000E+02, 2.17000E+02,
      2.23000E+02, 2.86760E+02,-2.93940E+00, 2.50000E+00, 0.00000E+00 };
  
-} // namespace
